+++ src/main/java/com/coinbase/prime/financing/FinancingService.java
@@ -304,7 +304,7 @@ public interface FinancingService {
    * @throws CoinbaseClientException if the request fails client-side validation
    * @throws CoinbasePrimeException if the Prime API returns an error response
    */
-  GetXmLiquidationResponse getXmLiquidation(GetXmLiquidationRequest request)
+  GetCrossMarginLiquidationResponse getCrossMarginLiquidation(GetCrossMarginLiquidationRequest request)
       throws CoinbaseClientException, CoinbasePrimeException;
 
   /**
@@ -317,7 +317,7 @@ public interface FinancingService {
    * @throws CoinbaseClientException if the request fails client-side validation
    * @throws CoinbasePrimeException if the Prime API returns an error response
    */
-  ListXmLiquidationsResponse listXmLiquidations(ListXmLiquidationsRequest request)
+  ListCrossMarginLiquidationsResponse listCrossMarginLiquidations(ListCrossMarginLiquidationsRequest request)
       throws CoinbaseClientException, CoinbasePrimeException;
 
   /**